refactor(theme): move duplicated theme switch tile code into ThemeSwitchTile widget in common/widgets/theme/
