+++ lib/mobile/features/settings/presentation/screens/settings_screen.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:streamkeys/common/widgets/forms/theme_mode_switch.dart';
+import 'package:streamkeys/common/widgets/theme/theme_switch_tile.dart';
 import 'package:streamkeys/core/app_update/presentation/widgets/app_version_tile.dart';
-import 'package:streamkeys/core/theme/bloc/theme_mode_bloc.dart';
 import 'package:streamkeys/mobile/features/api_connection/bloc/api_connection_bloc.dart';
 import 'package:streamkeys/mobile/features/api_connection/presentation/widgets/api_connection_tile.dart';
 import 'package:streamkeys/mobile/features/app_update/presentation/widgets/show_release_dialog.dart';
@@ -16,26 +15,7 @@ class SettingsScreen extends StatelessWidget {
     final content = SafeArea(
       child: Column(
         children: [
-          BlocBuilder<ThemeModeBloc, ThemeMode>(
-            builder: (context, themeMode) {
-              return ListTile(
-                title: const Text(
-                  'Theme mode',
-                  style: TextStyle(fontWeight: FontWeight.w500),
-                ),
-                trailing: Transform.scale(
-                  scale: 0.8,
-                  child: ThemeModeSwitch(
-                    themeMode: themeMode,
-                    onChanged: (themeMode) {
-                      final bloc = context.read<ThemeModeBloc>();
-                      bloc.add(ThemeModeChange(themeMode));
-                    },
-                  ),
-                ),
-              );
-            },
-          ),
+          const ThemeSwitchTile(),
           BlocBuilder<ApiConnectionBloc, ApiConnectionState>(
             builder: (context, state) {
               return ApiConnectionTile(
